Se creo la logica de los ServiceImpl

+++ src/main/java/com/torneos/futbol/repository/TorneoRepository.java
@@ -3,5 +3,5 @@ package com.torneos.futbol.repository;
 import com.torneos.futbol.model.entity.Torneo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TorneoRepository extends JpaRepository<Torneo, Long> {
+public interface TorneoRepository extends JpaRepository<Torneo, Integer> {
 }
